requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
tweepy>=4.14.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
playwright>=1.40.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
sentry-sdk>=1.38.0


